api/requirements/local.txt
-r common.txt
ruff
pre-commit
markdown
pytest-django
